/test: consolidate suites, docs + CLI parity, renewing arm for the PR snapshot error
Merge the auth transport and failure suites into the auth-service suite
and the routing/adoption/update-ladder suites into one composition suite.
Add the missing "renewing" arm to the PR snapshot auth error so a
mid-renewal state does not read as "auth is invalid". Give `ade github
app-auth status --text` a typed formatter that leads with credentialState
instead of the 8-hour expiry and prints the full refresh error. Update
the internal docs for the credential architecture, the honest status
model, and the storm-prevention design.
